+++ programs/drift/src/instructions/admin.rs
@@ -512,6 +512,13 @@ pub fn handle_update_serum_fulfillment_config_status(
     Ok(())
 }
 
+pub fn handle_delete_serum_fulfillment_config(
+    _ctx: Context<DeleteSerumFulfillmentConfig>,
+) -> Result<()> {
+    msg!("deleted serum fulfillment config");
+    Ok(())
+}
+
 pub fn handle_update_serum_vault(ctx: Context<UpdateSerumVault>) -> Result<()> {
     let vault = &ctx.accounts.srm_vault;
     validate!(
@@ -607,6 +614,13 @@ pub fn handle_update_openbook_v2_fulfillment_config_status(
     let mut config = load_mut!(ctx.accounts.openbook_v2_fulfillment_config)?;
     msg!("config.status {:?} -> {:?}", config.status, status);
     config.status = status;
+    Ok(())
+}
+
+pub fn handle_delete_openbook_v2_fulfillment_config(
+    _ctx: Context<DeleteOpenbookV2FulfillmentConfig>,
+) -> Result<()> {
+    msg!("deleted openbook v2 fulfillment config");
     Ok(())
 }
 
@@ -4484,19 +4498,19 @@ pub fn handle_update_spot_auction_duration(
     Ok(())
 }
 
-pub fn handle_admin_disable_update_perp_bid_ask_twap(
+pub fn handle_admin_update_user_stats_paused_operations(
     ctx: Context<AdminDisableBidAskTwapUpdate>,
-    disable: bool,
+    paused_operations: u8,
 ) -> Result<()> {
     let mut user_stats = load_mut!(ctx.accounts.user_stats)?;
 
     msg!(
-        "disable_update_perp_bid_ask_twap: {:?} -> {:?}",
-        user_stats.disable_update_perp_bid_ask_twap,
-        disable
+        "user_stats.paused_operations: {:?} -> {:?}",
+        user_stats.paused_operations,
+        paused_operations
     );
 
-    user_stats.disable_update_perp_bid_ask_twap = disable;
+    user_stats.paused_operations = paused_operations;
     Ok(())
 }
 
@@ -4856,7 +4870,6 @@ pub fn handle_admin_deposit<'c: 'info, 'info>(
         None,
     )?;
 
-    let user_token_amount_after = spot_position.get_signed_token_amount(&spot_market)?;
     let token_amount = spot_position.get_token_amount(&spot_market)?;
     if token_amount == 0 {
         validate!(
@@ -4881,6 +4894,7 @@ pub fn handle_admin_deposit<'c: 'info, 'info>(
     user.update_last_active_slot(slot);
 
     let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
+    let user_token_amount_after = user.get_total_token_amount(spot_market)?;
 
     controller::token::receive(
         &ctx.accounts.token_program,
@@ -5370,13 +5384,25 @@ pub struct InitializeSerumFulfillmentConfig<'info> {
 
 #[derive(Accounts)]
 pub struct UpdateSerumFulfillmentConfig<'info> {
-    #[account(
-        has_one = admin
-    )]
     pub state: Box<Account<'info, State>>,
     #[account(mut)]
     pub serum_fulfillment_config: AccountLoader<'info, SerumV3FulfillmentConfig>,
-    #[account(mut)]
+    #[account(
+        mut,
+        constraint = admin.key() == state.admin || admin.key() == admin_hot_wallet::id()
+    )]
+    pub admin: Signer<'info>,
+}
+
+#[derive(Accounts)]
+pub struct DeleteSerumFulfillmentConfig<'info> {
+    pub state: Box<Account<'info, State>>,
+    #[account(mut, close = admin)]
+    pub serum_fulfillment_config: AccountLoader<'info, SerumV3FulfillmentConfig>,
+    #[account(
+        mut,
+        constraint = admin.key() == state.admin || admin.key() == admin_hot_wallet::id()
+    )]
     pub admin: Signer<'info>,
 }
 
@@ -5949,13 +5975,25 @@ pub struct InitializeOpenbookV2FulfillmentConfig<'info> {
 
 #[derive(Accounts)]
 pub struct UpdateOpenbookV2FulfillmentConfig<'info> {
-    #[account(
-        has_one = admin
-    )]
     pub state: Box<Account<'info, State>>,
     #[account(mut)]
     pub openbook_v2_fulfillment_config: AccountLoader<'info, OpenbookV2FulfillmentConfig>,
-    #[account(mut)]
+    #[account(
+        mut,
+        constraint = admin.key() == state.admin || admin.key() == admin_hot_wallet::id()
+    )]
+    pub admin: Signer<'info>,
+}
+
+#[derive(Accounts)]
+pub struct DeleteOpenbookV2FulfillmentConfig<'info> {
+    pub state: Box<Account<'info, State>>,
+    #[account(mut, close = admin)]
+    pub openbook_v2_fulfillment_config: AccountLoader<'info, OpenbookV2FulfillmentConfig>,
+    #[account(
+        mut,
+        constraint = admin.key() == state.admin || admin.key() == admin_hot_wallet::id()
+    )]
     pub admin: Signer<'info>,
 }