+++ sdk/src/types.ts
@@ -80,6 +80,12 @@ export enum UserStatus {
 	PROTECTED_MAKER = 16,
 }
 
+export enum UserStatsPausedOperation {
+	UPDATE_BID_ASK_TWAP = 1,
+	AMM_ATOMIC_FILL = 2,
+	AMM_ATOMIC_RISK_INCREASING_FILL = 4,
+}
+
 export class MarginMode {
 	static readonly DEFAULT = { default: {} };
 	static readonly HIGH_LEVERAGE = { highLeverage: {} };
@@ -1134,7 +1140,6 @@ export type PerpPosition = {
 	/**	 TODO: remove this field - it doesn't exist on chain */
 	remainderBaseAssetAmount: number;
 	maxMarginRatio: number;
-	lastBaseAssetAmountPerLp: BN;
 	lastQuoteAssetAmountPerLp: BN;
 	perLpBase: number;
 	isolatedPositionScaledBalance: BN;