+++ CHANGELOG.md
@@ -9,6 +9,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: isolated positions [#1757](https://github.com/drift-labs/protocol-v2/pull/1757)
+- program: delete serum/openbook configs [#2066](https://github.com/drift-labs/protocol-v2/pull/2066)
+
+### Fixes
+
+### Breaking
+
+## [2.152.0] - 2025-12-12
+
+### Features
+
+- program: add bit_flags to liquidation record [#2053](https://github.com/drift-labs/protocol-v2/pull/2053)
+
+### Fixes
+
+- program: fix base spread validate for perp market [#2052](https://github.com/drift-labs/protocol-v2/pull/2052)
+- sdk: fix for initializing a new user account with a token2022 deposit
+
+### Breaking
+
+## [2.151.0] - 2025-12-03
+
+### Features
+
+- program: add reduce only atomic taking against amm option [#2037](https://github.com/drift-labs/protocol-v2/pull/2037)
+
+### Fixes
+
+### Breaking
+
+## [2.150.0] - 2025-12-01
+
+### Features
+
+- program: stricter logic for atomic fills [#2042](https://github.com/drift-labs/protocol-v2/pull/2042)
+- program: reset lp fields in update_user_idle [#2018](https://github.com/drift-labs/protocol-v2/pull/2018)
+
 ### Fixes
 
 ### Breaking